requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
alembic==1.12.0
annotated-types==0.5.0
anyio==4.0.0
astroid==2.15.6
async-timeout==4.0.3
attrs==23.1.0
Authlib==1.2.1
bcrypt==4.0.1
black==23.11.0
boto3==1.28.38
botocore==1.31.38
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
coverage==7.3.2
cryptography==41.0.3
dill==0.3.7
dnspython==2.4.2
ecdsa==0.18.0
email-validator==2.0.0.post2
fastapi==0.103.0
ffmpeg-python==0.2.0
filelock==3.12.3
flake8==6.1.0
flake8-isort==6.1.1
frozenlist==1.4.0
future==0.18.3
GeoAlchemy2==0.14.1
greenlet==2.0.2
gunicorn==21.2.0
h11==0.14.0
hiredis==2.2.3
httpcore==1.0.2
httptools==0.6.0
httpx==0.25.1
idna==3.4
iniconfig==2.0.0
isort==5.12.0
itsdangerous==2.1.2
jedi==0.19.0
Jinja2==3.1.2
jmespath==1.0.1
lazy-object-proxy==1.9.0
makefun==1.15.1
Mako==1.2.4
MarkupSafe==2.1.3
mccabe==0.7.0
more-itertools==10.1.0
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
networkx==3.1
packaging==23.1
parso==0.8.3
passlib==1.7.4
pathspec==0.11.2
pika==1.3.2
platformdirs==3.10.0
pluggy==1.3.0
psycopg2==2.9.7
pyasn1==0.5.0
pycodestyle==2.11.1
pycparser==2.21
pydantic==2.3.0
pydantic-settings==2.0.3
pydantic_core==2.6.3
pyflakes==3.1.0
PyJWT==2.8.0
pylint==2.17.5
pypng==0.20220715.0
pyscopg2==66.0.2
pytest==7.4.2
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
PyYAML==6.0.1
qrcode==7.4.2
redis==5.0.0
regex==2023.8.8
requests==2.31.0
rsa==4.9
s3transfer==0.6.2
semantic-version==2.10.0
setuptools-rust==1.7.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.20
starlette==0.27.0
starlette-authlib==0.1.30
sympy==1.12
tiktoken==0.3.1
tomlkit==0.12.1
tqdm==4.66.1
typing_extensions==4.7.1
urllib3==1.26.16
uvicorn==0.23.2
uvloop==0.17.0
watchfiles==0.20.0
Wave==0.0.2
websockets==11.0.3
wrapt==1.15.0
yarl==1.9.2
